moved decorators outside of openai_service
